refactor plugins.py so that functionality is outsourced to plugin_utils.py where possible
