+++ ckanext/switzerland/plugins.py
@@ -3,23 +3,18 @@
 from ckanext.switzerland import validators as v
 from ckanext.switzerland import logic as l
 import ckanext.switzerland.helpers as sh
-
+import ckanext.switzerland.plugin_utils as pu
+import re
+from webhelpers.html import HTML
+from webhelpers import paginate
 import ckan.plugins as plugins
 from ckan.lib.plugins import DefaultTranslation
 import ckanext.xloader.interfaces as ix
 import ckan.plugins.toolkit as toolkit
-from ckan import logic
 import ckan.lib.helpers as h
-from ckan.lib.munge import munge_title_to_name
-import json
-import re
 import collections
-from webhelpers.html import HTML
-from webhelpers import paginate
-import urlparse
 import os
 import logging
-import yaml
 log = logging.getLogger(__name__)
 
 __location__ = os.path.realpath(os.path.join(
@@ -140,306 +135,64 @@ class OgdchPlugin(plugins.SingletonPlugin, DefaultTranslation):
         }
 
 
-class FormatMappingNotLoadedError(Exception):
-    pass
-
-
-class OgdchLanguagePlugin(plugins.SingletonPlugin):
+class OgdchMixin(object):
     """
-    Handles language dictionaries in data_dict (pkg_dict).
+    gets format mapping
     """
-    plugins.implements(plugins.IConfigurer)
-
-    # IConfigurer
-
     def update_config(self, config):
-        try:
-            mapping_path = os.path.join(__location__, 'mapping.yaml')
-            with open(mapping_path, 'r') as format_mapping_file:
-                self.format_mapping = yaml.safe_load(format_mapping_file)
-        except (IOError, yaml.YAMLError) as exception:
-            raise FormatMappingNotLoadedError(
-                'Loading Format-Mapping from Path: (%s) '
-                'failed with Exception: (%s)'
-                % (mapping_path, exception)
-            )
-
-    def get_format_mapping(self):
-        return self.format_mapping
-
-    def before_view(self, pkg_dict):
-        pkg_dict = self._prepare_package_json(pkg_dict)
-
-        return pkg_dict
-
-    def _ignore_field(self, key):
-        return False
-
-    def _prepare_package_json(self, pkg_dict):
-        # parse all json strings in dict
-        pkg_dict = self._package_parse_json_strings(pkg_dict)
-
-        # map ckan fields
-        pkg_dict = self._package_map_ckan_default_fields(pkg_dict)
-
-        # prepare format of resources
-        pkg_dict = self._prepare_resources_format(pkg_dict)
-
-        try:
-            # Do not change the resulting dict for API requests
-            path = toolkit.request.path
-            if any([
-                path.startswith('/api'),
-                path.endswith('.xml'),
-                path.endswith('.rdf'),
-                path.endswith('.n3'),
-                path.endswith('.ttl'),
-                path.endswith('.jsonld'),
-
-            ]):
-                return pkg_dict
-        except TypeError:
-            # we get here if there is no request (i.e. on the command line)
-            return pkg_dict
-
-        # replace langauge dicts with requested language strings
-        desired_lang_code = self._get_request_language()
-        pkg_dict = self._package_reduce_to_requested_language(
-            pkg_dict, desired_lang_code
-        )
-
-        return pkg_dict
-
-    def _get_request_language(self):
-        try:
-            return toolkit.request.environ['CKAN_LANG']
-        except TypeError:
-            return toolkit.config.get('ckan.locale_default', 'en')
-
-    def _package_parse_json_strings(self, pkg_dict):
-        # try to parse all values as JSON
-        for key, value in pkg_dict.iteritems():
-            pkg_dict[key] = sh.parse_json(value)
-
-        # groups
-        if 'groups' in pkg_dict and pkg_dict['groups'] is not None:
-            for group in pkg_dict['groups']:
-                """
-                TODO: somehow the title is messed up here,
-                but the display_name is okay
-                """
-                group['title'] = group['display_name']
-                for field in group:
-                    group[field] = sh.parse_json(group[field])
-
-        # organization
-        if 'organization' in pkg_dict and pkg_dict['organization'] is not None:
-            for field in pkg_dict['organization']:
-                pkg_dict['organization'][field] = sh.parse_json(
-                    pkg_dict['organization'][field]
-                )
-
-        return pkg_dict
-
-    def _package_map_ckan_default_fields(self, pkg_dict):  # noqa
-        pkg_dict['display_name'] = pkg_dict['title']
-
-        if pkg_dict.get('maintainer') is None:
-            try:
-                pkg_dict['maintainer'] = pkg_dict['contact_points'][0]['name']  # noqa
-            except (KeyError, IndexError):
-                pass
-
-        if pkg_dict.get('maintainer_email') is None:
-            try:
-                pkg_dict['maintainer_email'] = pkg_dict['contact_points'][0]['email']  # noqa
-            except (KeyError, IndexError):
-                pass
-
-        if pkg_dict.get('author') is None:
-            try:
-                pkg_dict['author'] = pkg_dict['publishers'][0]['label']  # noqa
-            except (KeyError, IndexError):
-                pass
-
-        if pkg_dict.get('resources') is not None:
-            for resource in pkg_dict['resources']:
-                resource['name'] = resource['title']
-
-        return pkg_dict
-
-    def _prepare_resources_format(self, pkg_dict):
-        if pkg_dict.get('resources') is not None:
-            for resource in pkg_dict['resources']:
-                resource = self._prepare_resource_format(resource)
-
-                # if format could not be mapped and media_type exists use this value  # noqa
-                if (not resource.get('format') and resource.get('media_type')):
-                    resource['format'] = resource['media_type'].split('/')[-1]
-
-        return pkg_dict
-
-    # Generates format of resource and saves it in format field
-    def _prepare_resource_format(self, resource):
-        resource_format = ''
-
-        # get format from media_type field if available
-        if not resource_format and resource.get('media_type'):  # noqa
-            resource_format = resource['media_type'].split('/')[-1].lower()
-
-        # get format from format field if available (lol)
-        if not resource_format and resource.get('format'):
-            resource_format = resource['format'].split('/')[-1].lower()
-
-        # check if 'media_type' or 'format' can be mapped
-        has_format = (sh.map_to_valid_format(
-            resource_format,
-            self.get_format_mapping()
-        ) is not None)
-
-        # if the fields can't be mapped,
-        # try to parse the download_url as a last resort
-        if not has_format and resource.get('download_url'):
-            path = urlparse.urlparse(resource['download_url']).path
-            ext = os.path.splitext(path)[1]
-            if ext:
-                resource_format = ext.replace('.', '').lower()
-
-        mapped_format = sh.map_to_valid_format(
-            resource_format,
-            self.get_format_mapping()
-        )
-        if mapped_format:
-            # if format could be successfully mapped write it to format field
-            resource['format'] = mapped_format
-        elif not resource.get('download_url'):
-            resource['format'] = 'SERVICE'
-        else:
-            # else return empty string (this will be indexed as N/A)
-            resource['format'] = ''
-
-        return resource
-
-    def _extract_lang_value(self, value, lang_code):
-        new_value = sh.parse_json(value)
-
-        if isinstance(new_value, dict):
-            return sh.get_localized_value(
-                new_value,
-                lang_code,
-                default_value=''
-            )
-        return value
-
-    def _package_reduce_to_requested_language(self, pkg_dict, desired_lang_code):  # noqa
-        # pkg fields
-        for key, value in pkg_dict.iteritems():
-            if not self._ignore_field(key):
-                pkg_dict[key] = self._extract_lang_value(
-                    value,
-                    desired_lang_code
-                )
-
-        # groups
-        pkg_dict = self._reduce_group_language(pkg_dict, desired_lang_code)
-
-        # organization
-        pkg_dict = self._reduce_org_language(pkg_dict, desired_lang_code)
-
-        # resources
-        pkg_dict = self._reduce_res_language(pkg_dict, desired_lang_code)
-
-        return pkg_dict
-
-    def _reduce_group_language(self, pkg_dict, desired_lang_code):
-        if 'groups' in pkg_dict and pkg_dict['groups'] is not None:
-            try:
-                for element in pkg_dict['groups']:
-                    for field in element:
-                        element[field] = self._extract_lang_value(
-                            element[field],
-                            desired_lang_code
-                        )
-            except TypeError:
-                pass
-
-        return pkg_dict
-
-    def _reduce_org_language(self, pkg_dict, desired_lang_code):
-        if 'organization' in pkg_dict and pkg_dict['organization'] is not None:
-            try:
-                for field in pkg_dict['organization']:
-                    pkg_dict['organization'][field] = self._extract_lang_value(
-                        pkg_dict['organization'][field],
-                        desired_lang_code
-                    )
-            except TypeError:
-                pass
-        return pkg_dict
-
-    def _reduce_res_language(self, pkg_dict, desired_lang_code):
-        if 'resources' in pkg_dict and pkg_dict['resources'] is not None:
-            try:
-                for element in pkg_dict['resources']:
-                    for field in element:
-                        element[field] = self._extract_lang_value(
-                            element[field],
-                            desired_lang_code
-                        )
-            except TypeError:
-                pass
-        return pkg_dict
+        self.format_mapping = \
+            pu.ogdch_get_format_mapping()
 
 
-class OgdchGroupPlugin(OgdchLanguagePlugin):
+class OgdchGroupPlugin(plugins.SingletonPlugin, OgdchMixin):
+    plugins.implements(plugins.IConfigurer, inherit=True)
     plugins.implements(plugins.IGroupController, inherit=True)
 
-    # IGroupController
     def before_view(self, pkg_dict):
-        return super(OgdchGroupPlugin, self).before_view(pkg_dict)
+        pkg_dict = pu._prepare_package_json(
+            pkg_dict=pkg_dict,
+            format_mapping=self.format_mapping,
+            ignore_fields=[]
+        )
+        return pkg_dict
 
 
-class OgdchOrganizationPlugin(OgdchLanguagePlugin):
+class OgdchOrganizationPlugin(plugins.SingletonPlugin, OgdchMixin):
+    plugins.implements(plugins.IConfigurer, inherit=True)
     plugins.implements(plugins.IOrganizationController, inherit=True)
 
-    # IOrganizationController
     def before_view(self, pkg_dict):
-        return super(OgdchOrganizationPlugin, self).before_view(pkg_dict)
+        pkg_dict = pu._prepare_package_json(
+            pkg_dict=pkg_dict,
+            format_mapping=self.format_mapping,
+            ignore_fields=[]
+        )
+        return pkg_dict
 
 
-class OgdchResourcePlugin(OgdchLanguagePlugin):
+class OgdchResourcePlugin(plugins.SingletonPlugin, OgdchMixin):
+    plugins.implements(plugins.IConfigurer, inherit=True)
     plugins.implements(plugins.IResourceController, inherit=True)
 
     # IResourceController
     def before_show(self, res_dict):
-        res_dict = super(OgdchResourcePlugin, self).before_view(res_dict)
-        res_dict = self._prepare_resource_format(res_dict)
-
-        # if format could not be mapped and media_type exists use this value
-        if not res_dict.get('format') and res_dict.get('media_type'):
-            res_dict['format'] = res_dict['media_type'].split('/')[-1]
-
-        return res_dict
-
-    def _ignore_field(self, key):
-        return key == 'tracking_summary'
+        pu.ogdch_prepare_res_dict_before_show(
+            res_dict=res_dict,
+            format_mapping=self.format_mapping,
+            ignore_fields=['tracking_summary']
+        )
 
 
-class OgdchPackagePlugin(OgdchLanguagePlugin):
+class OgdchPackagePlugin(plugins.SingletonPlugin, OgdchMixin):
+    plugins.implements(plugins.IConfigurer, inherit=True)
     plugins.implements(plugins.IPackageController, inherit=True)
     plugins.implements(plugins.IRoutes, inherit=True)
     plugins.implements(ix.IXloader, inherit=True)
 
-    def is_supported_package_type(self, pkg_dict):
-        # only package type 'dataset' is supported (not harvesters!)
-        try:
-            return (pkg_dict['type'] == 'dataset')
-        except KeyError:
-            return False
-
     # IRouter
-    # create perma-link route
+
     def before_map(self, map):
+        """create perma-link route"""
         map.connect('perma_redirect', '/perma/{id}',
                     controller='ckanext.switzerland.controller:OgdchPermaController',  # noqa
                     action='read')
@@ -447,216 +200,40 @@ class OgdchPackagePlugin(OgdchLanguagePlugin):
 
     # IPackageController
 
-#     TODO: before_view isn't called in API requests -> after_show is
-#           BUT (!) after_show is also called when packages get indexed
-#           and there we need all languages.
-#           -> find a solution to _prepare_package_json() in an API call.
-#     def after_show(self, context, pkg_dict):
-#         if not self.is_supported_package_type(pkg_dict):
-#             return pkg_dict
-#
-#         return super(OgdchPackagePlugin, self).before_view(pkg_dict)
+    def before_view(self, pkg_dict):
+        """transform pkg dict before view"""
+        pkg_dict = pu._prepare_package_json(
+            pkg_dict=pkg_dict,
+            format_mapping=self.format_mapping,
+            ignore_fields=[]
+        )
+        return pkg_dict
 
     def after_show(self, context, pkg_dict):
-        if not self.is_supported_package_type(pkg_dict):
-            return pkg_dict
-
-        pkg_dict = self._package_map_ckan_default_fields(pkg_dict)
-
-        # groups
-        if pkg_dict['groups'] is not None:
-            for group in pkg_dict['groups']:
-                """
-                TODO: somehow the title is messed up here,
-                but the display_name is okay
-                """
-                group['title'] = group['display_name']
-                for field in group:
-                    group[field] = sh.parse_json(group[field])
-
-        # load organization from API to get all fields defined in schema
-        # by default, CKAN loads organizations only from the database
-        if pkg_dict['owner_org'] is not None:
-            pkg_dict['organization'] = logic.get_action('organization_show')(
-                {},
-                {
-                    'id': pkg_dict['owner_org'],
-                    'include_users': False,
-                    'include_followers': False,
-                }
-            )
-
+        """
+        before_view isn't called in API requests -> after_show is
+        BUT (!) after_show is also called when packages get indexed
+        and there we need all languages.
+        -> find a solution to _prepare_package_json() in an API call.
+        """
+        pkg_dict = pu.ogdch_prepare_pkg_dict_for_api(pkg_dict)
         return pkg_dict
 
     def before_index(self, search_data):
-        if not self.is_supported_package_type(search_data):
-            return search_data
-
-        extract_title = LangToString('title')
-        validated_dict = json.loads(search_data['validated_data_dict'])
-
-        search_data['res_name'] = [extract_title(r) for r in validated_dict[u'resources']]  # noqa
-        search_data['res_name_en'] = [sh.get_localized_value(r['title'], 'en') for r in validated_dict[u'resources']]  # noqa
-        search_data['res_name_de'] = [sh.get_localized_value(r['title'], 'de') for r in validated_dict[u'resources']]  # noqa
-        search_data['res_name_fr'] = [sh.get_localized_value(r['title'], 'fr') for r in validated_dict[u'resources']]  # noqa
-        search_data['res_name_it'] = [sh.get_localized_value(r['title'], 'it') for r in validated_dict[u'resources']]  # noqa
-        search_data['res_description_en'] = [sh.get_localized_value(r['description'], 'en') for r in validated_dict[u'resources']]  # noqa
-        search_data['res_description_de'] = [sh.get_localized_value(r['description'], 'de') for r in validated_dict[u'resources']]  # noqa
-        search_data['res_description_fr'] = [sh.get_localized_value(r['description'], 'fr') for r in validated_dict[u'resources']]  # noqa
-        search_data['res_description_it'] = [sh.get_localized_value(r['description'], 'it') for r in validated_dict[u'resources']]  # noqa
-        search_data['groups_en'] = [sh.get_localized_value(g['display_name'], 'en') for g in validated_dict[u'groups']]  # noqa
-        search_data['groups_de'] = [sh.get_localized_value(g['display_name'], 'de') for g in validated_dict[u'groups']]  # noqa
-        search_data['groups_fr'] = [sh.get_localized_value(g['display_name'], 'fr') for g in validated_dict[u'groups']]  # noqa
-        search_data['groups_it'] = [sh.get_localized_value(g['display_name'], 'it') for g in validated_dict[u'groups']]  # noqa
-        search_data['res_description'] = [LangToString('description')(r) for r in validated_dict[u'resources']]  # noqa
-        search_data['res_format'] = self._prepare_formats_for_index(validated_dict[u'resources'])  # noqa
-        search_data['res_rights'] = [sh.simplify_terms_of_use(r['rights']) for r in validated_dict[u'resources'] if 'rights' in r.keys()]  # noqa
-        search_data['title_string'] = extract_title(validated_dict)
-        search_data['description'] = LangToString('description')(validated_dict)  # noqa
-        if 'political_level' in validated_dict[u'organization']:
-            search_data['political_level'] = validated_dict[u'organization'][u'political_level']  # noqa
-
-        search_data['identifier'] = validated_dict.get('identifier')
-        search_data['contact_points'] = [c['name'] for c in validated_dict.get('contact_points', [])]  # noqa
-        search_data['publishers'] = [p['label'] for p in validated_dict.get('publishers', [])]  # noqa
-
-        # TODO: Remove the try-except-block.
-        # This fixes the index while we have 'wrong' relations on
-        # datasets harvested with an old version of ckanext-geocat
-        try:
-            search_data['see_alsos'] = [d['dataset_identifier'] for d in validated_dict.get('see_alsos', [])]  # noqa
-        except TypeError:
-            search_data['see_alsos'] = [d for d in
-                                        validated_dict.get('see_alsos',
-                                                           [])]  # noqa
-
-        # make sure we're not dealing with NoneType
-        if search_data['metadata_created'] is None:
-            search_data['metadata_created'] = ''
-
-        if search_data['metadata_modified'] is None:
-            search_data['metadata_modified'] = ''
-
-        try:
-            # index language-specific values (or it's fallback)
-            for lang_code in sh.get_langs():
-                search_data['title_' + lang_code] = sh.get_localized_value(
-                    validated_dict['title'],
-                    lang_code
-                )
-                search_data['title_string_' + lang_code] = munge_title_to_name(
-                    sh.get_localized_value(validated_dict['title'], lang_code)
-                )
-                search_data['description_' + lang_code] = sh.get_localized_value(  # noqa
-                    validated_dict['description'],
-                    lang_code
-                )
-                search_data['keywords_' + lang_code] = sh.get_localized_value(
-                    validated_dict['keywords'],
-                    lang_code
-                )
-                search_data['organization_' + lang_code] = sh.get_localized_value(  # noqa
-                    validated_dict['organization']['title'],
-                    lang_code
-                )
-
-        except KeyError:
-            pass
-
-        # clean terms for suggest context
-        search_data = self._prepare_suggest_context(
-            search_data,
-            validated_dict
+        """
+        Search data before index
+        """
+        search_data = pu.ogdch_prepare_search_data_for_index(
+            search_data=search_data,
+            format_mapping=self.format_mapping
         )
-
         return search_data
 
-    # generates a set with formats of all resources
-    def _prepare_formats_for_index(self, resources):
-        formats = set()
-        for r in resources:
-            resource = self._prepare_resource_format(r)
-            if resource['format']:
-                formats.add(resource['format'])
-            else:
-                formats.add('N/A')
-
-        return list(formats)
-
-    def _prepare_suggest_context(self, search_data, pkg_dict):
-        def clean_suggestion(term):
-            return term.replace('-', '')
-
-        search_data['suggest_groups'] = [clean_suggestion(t['name']) for t in pkg_dict['groups']]  # noqa
-        search_data['suggest_organization'] = clean_suggestion(pkg_dict['organization']['name'])  # noqa
-
-        search_data['suggest_tags'] = []
-        search_data['suggest_tags'].extend([clean_suggestion(t) for t in search_data.get('keywords_de', [])])  # noqa
-        search_data['suggest_tags'].extend([clean_suggestion(t) for t in search_data.get('keywords_fr', [])])  # noqa
-        search_data['suggest_tags'].extend([clean_suggestion(t) for t in search_data.get('keywords_it', [])])  # noqa
-        search_data['suggest_tags'].extend([clean_suggestion(t) for t in search_data.get('keywords_en', [])])  # noqa
-
-        search_data['suggest_res_rights'] = [clean_suggestion(t) for t in search_data['res_rights']]  # noqa
-        search_data['suggest_res_format'] = [clean_suggestion(t) for t in search_data['res_format']]  # noqa
-
-        return search_data
-
-    # borrowed from ckanext-multilingual (core extension)
     def before_search(self, search_params):
         """
         Adjust search parameters
         """
-
-        '''
-        search in correct language-specific field and boost
-        results in current language
-        '''
-        lang_set = sh.get_langs()
-        try:
-            current_lang = toolkit.request.environ['CKAN_LANG']
-        except TypeError as err:
-            if err.message == ('No object (name: request) has been registered '
-                               'for this thread'):
-                # This happens when this code gets called as part of a paster
-                # command rather then as part of an HTTP request.
-                current_lang = toolkit.config.get('ckan.locale_default')
-            else:
-                raise
-
-        # fallback to default locale if locale not in suported langs
-        if current_lang not in lang_set:
-            current_lang = toolkit.config.get('ckan.locale_default', 'en')
-        # treat current lang differenly so remove from set
-        lang_set.remove(current_lang)
-
-        # add default query field(s)
-        query_fields = 'text'
-
-        # weight current lang more highly
-        query_fields += ' title_%s^8 text_%s^4' % (current_lang, current_lang)
-
-        for lang in lang_set:
-            query_fields += ' title_%s^2 text_%s' % (lang, lang)
-
-        search_params['qf'] = query_fields
-
-        '''
-        Unless the query is already being filtered by any type
-        (either positively, or negatively), reduce to only display
-        'dataset' type
-        This is done because by standard all types are displayed, this
-        leads to strange situations where e.g. harvest sources are shown
-        on organization pages.
-        TODO: fix issue https://github.com/ckan/ckan/issues/2803 in CKAN core
-        '''
-        fq = search_params.get('fq', '')
-        if 'dataset_type:' not in fq:
-            search_params.update({'fq': "%s +dataset_type:dataset" % fq})
-
-        # remove colon followed by a space from q to avoid false negatives
-        q = search_params.get('q', '')
-        search_params['q'] = re.sub(":\s", " ", q)
-
+        search_params = pu.ogdch_adjust_search_params(search_params)
         return search_params
 
     # IXloader
@@ -701,25 +278,6 @@ class OgdchGroupSearchPlugin(plugins.SingletonPlugin):
                     controller='ckanext.switzerland.controller:OgdchGroupSearchController',  # noqa
                     action='read')
         return map
-
-
-class LangToString(object):
-    def __init__(self, attribute):
-        self.attribute = attribute
-
-    def __call__(self, data_dict):
-        try:
-            lang = data_dict[self.attribute]
-            return (
-                '%s - %s - %s - %s' % (
-                    lang.get('de', ''),
-                    lang.get('fr', ''),
-                    lang.get('it', ''),
-                    lang.get('en', '')
-                )
-            )
-        except KeyError:
-            return ''
 
 
 # Monkeypatch to style CKAN pagination
