+++ ckanext/switzerland/tests/test_plugin.py
@@ -1,7 +1,7 @@
 """Tests for plugin.py."""
-import ckanext.switzerland.plugins as plugin
+import ckanext.switzerland.plugins as plugins
+import ckanext.switzerland.plugin_utils as pu
 from nose.tools import *  # noqa
-import mock
 import sys
 import os
 import yaml
@@ -29,7 +29,7 @@ class TestPlugin(unittest.TestCase):
             return yaml.safe_load(format_mapping_file)
 
     def test_prepare_resource_format(self):
-        ogdch_language_plugin = plugin.OgdchLanguagePlugin()
+        ogdch_language_plugin = plugins.OgdchMixin()
 
         ogdch_language_plugin.format_mapping = self._load_test_format_mapping()
 
@@ -38,7 +38,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': None,
             'format': None
         }
-        resource_with_no_format_and_no_download_url_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_no_format_and_no_download_url.copy())
+        resource_with_no_format_and_no_download_url_cleaned = pu._prepare_resource_format(
+            resource_with_no_format_and_no_download_url.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('SERVICE', resource_with_no_format_and_no_download_url_cleaned['format'])
 
         resource_with_invalid_format_and_with_download_url = {
@@ -46,7 +47,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': None,
             'format': 'dogvideo'
         }
-        resource_with_invalid_format_and_with_download_url_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_invalid_format_and_with_download_url.copy())
+        resource_with_invalid_format_and_with_download_url_cleaned = pu._prepare_resource_format(
+            resource_with_invalid_format_and_with_download_url.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('', resource_with_invalid_format_and_with_download_url_cleaned['format'])
 
         resource_without_any_formats_with_download_url = {
@@ -54,7 +56,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': None,
             'format': None
         }
-        resource_without_any_formats_with_download_url_cleaned = ogdch_language_plugin._prepare_resource_format(resource_without_any_formats_with_download_url.copy())
+        resource_without_any_formats_with_download_url_cleaned = pu._prepare_resource_format(
+            resource_without_any_formats_with_download_url.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('', resource_without_any_formats_with_download_url_cleaned['format'])
 
         resource_with_invalid_format_and_no_download_url = {
@@ -62,7 +65,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': None,
             'format': 'catgif'
         }
-        resource_with_invalid_format_and_no_download_url_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_invalid_format_and_no_download_url.copy())
+        resource_with_invalid_format_and_no_download_url_cleaned = pu._prepare_resource_format(
+            resource_with_invalid_format_and_no_download_url.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('SERVICE', resource_with_invalid_format_and_no_download_url_cleaned['format'])
 
         resource_with_valid_format = {
@@ -70,7 +74,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': None,
             'format': 'xml'
         }
-        resource_with_valid_format_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_valid_format.copy())
+        resource_with_valid_format_cleaned = pu._prepare_resource_format(
+            resource_with_valid_format.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('XML', resource_with_valid_format_cleaned['format'])
 
         resource_with_invalid_media_type = {
@@ -78,7 +83,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': 'cat/gif',
             'format': 'gif'
         }
-        resource_with_invalid_media_type_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_invalid_media_type.copy())
+        resource_with_invalid_media_type_cleaned = pu._prepare_resource_format(
+            resource_with_invalid_media_type.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('', resource_with_invalid_media_type_cleaned['format'])
 
         resource_with_valid_media_type_without_slash = {
@@ -86,7 +92,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': 'html',
             'format': 'xml'
         }
-        resource_with_valid_media_type_without_slash_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_valid_media_type_without_slash.copy())
+        resource_with_valid_media_type_without_slash_cleaned = pu._prepare_resource_format(
+            resource_with_valid_media_type_without_slash.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('HTML', resource_with_valid_media_type_without_slash_cleaned['format'])
 
         resource_with_valid_media_type_with_slash = {
@@ -94,7 +101,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': 'text/html',
             'format': 'xml'
         }
-        resource_with_valid_media_type_with_slash_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_valid_media_type_with_slash.copy())
+        resource_with_valid_media_type_with_slash_cleaned = pu._prepare_resource_format(
+            resource_with_valid_media_type_with_slash.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('HTML', resource_with_valid_media_type_with_slash_cleaned['format'])
 
         resourse_with_download_url_without_extension = {
@@ -102,7 +110,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': 'text/xml',
             'format': 'html'
         }
-        resourse_with_download_url_without_extension_cleaned = ogdch_language_plugin._prepare_resource_format(resourse_with_download_url_without_extension.copy())
+        resourse_with_download_url_without_extension_cleaned = pu._prepare_resource_format(
+            resourse_with_download_url_without_extension.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('XML', resourse_with_download_url_without_extension_cleaned['format'])
 
         resourse_with_download_url_with_invalid_extension = {
@@ -110,7 +119,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': '',
             'format': ''
         }
-        resourse_with_download_url_with_invalid_extension_cleaned = ogdch_language_plugin._prepare_resource_format(resourse_with_download_url_with_invalid_extension.copy())
+        resourse_with_download_url_with_invalid_extension_cleaned = pu._prepare_resource_format(
+            resourse_with_download_url_with_invalid_extension.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('', resourse_with_download_url_with_invalid_extension_cleaned['format'])
 
         resourse_with_download_url_with_invalid_extension_but_format = {
@@ -118,7 +128,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': 'text/xml',
             'format': 'xml'
         }
-        resourse_with_download_url_with_invalid_extension_but_format_cleaned = ogdch_language_plugin._prepare_resource_format(resourse_with_download_url_with_invalid_extension_but_format.copy())
+        resourse_with_download_url_with_invalid_extension_but_format_cleaned = pu._prepare_resource_format(
+            resourse_with_download_url_with_invalid_extension_but_format.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('XML', resourse_with_download_url_with_invalid_extension_but_format_cleaned['format'])
 
         resourse_with_download_url_with_valid_extension = {
@@ -126,7 +137,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': '',
             'format': ''
         }
-        resourse_with_download_url_with_valid_extension_cleaned = ogdch_language_plugin._prepare_resource_format(resourse_with_download_url_with_valid_extension.copy())
+        resourse_with_download_url_with_valid_extension_cleaned = pu._prepare_resource_format(
+            resourse_with_download_url_with_valid_extension.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('ZIP', resourse_with_download_url_with_valid_extension_cleaned['format'])
 
         resourse_with_download_url_with_valid_extension_but_format = {
@@ -134,7 +146,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': 'text/xml',
             'format': 'xml'
         }
-        resourse_with_download_url_with_valid_extension_but_format_cleaned = ogdch_language_plugin._prepare_resource_format(resourse_with_download_url_with_valid_extension_but_format.copy())
+        resourse_with_download_url_with_valid_extension_but_format_cleaned = pu._prepare_resource_format(
+            resourse_with_download_url_with_valid_extension_but_format.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('XML', resourse_with_download_url_with_valid_extension_but_format_cleaned['format'])
 
         resource_with_ods_vndoas_format = {
@@ -142,7 +155,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': None,
             'format': 'application/vnd.oas...'
         }
-        resource_with_ods_vndoas_format_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_ods_vndoas_format.copy())
+        resource_with_ods_vndoas_format_cleaned = pu._prepare_resource_format(
+            resource_with_ods_vndoas_format.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('ODS', resource_with_ods_vndoas_format_cleaned['format'])
 
         resource_with_vndoxml_format = {
@@ -150,7 +164,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': None,
             'format': 'vnd.openxmlformats-officedocument.spreadsheetml.sheet'
         }
-        resource_with_vndoxml_format_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_vndoxml_format.copy())
+        resource_with_vndoxml_format_cleaned = pu._prepare_resource_format(
+            resource_with_vndoxml_format.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('XLS', resource_with_vndoxml_format_cleaned['format'])
 
         resource_with_pcaxis_format = {
@@ -158,7 +173,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': None,
             'format': 'pc-axis file'
         }
-        resource_with_pcaxis_format_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_pcaxis_format.copy())
+        resource_with_pcaxis_format_cleaned = pu._prepare_resource_format(
+            resource_with_pcaxis_format.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('PC-AXIS', resource_with_pcaxis_format_cleaned['format'])
 
         resource_with_pcaxis_format = {
@@ -166,7 +182,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': 'pc-axis file',
             'format': 'CSV'
         }
-        resource_with_pcaxis_format_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_pcaxis_format.copy())
+        resource_with_pcaxis_format_cleaned = pu._prepare_resource_format(
+            resource_with_pcaxis_format.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('PC-AXIS', resource_with_pcaxis_format_cleaned['format'])
 
         resource_with_rdf_sparql_format = {
@@ -174,7 +191,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': None,
             'format': 'Application/Sparql-...'
         }
-        resource_with_rdf_sparql_format_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_rdf_sparql_format.copy())
+        resource_with_rdf_sparql_format_cleaned = pu._prepare_resource_format(
+            resource_with_rdf_sparql_format.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('RDF', resource_with_rdf_sparql_format_cleaned['format'])
 
         resource_with_shapefile_esri_format = {
@@ -182,7 +200,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': None,
             'format': 'ESRI Shapefile'
         }
-        resource_with_shapefile_esri_format_cleaned = ogdch_language_plugin._prepare_resource_format(resource_with_shapefile_esri_format.copy())
+        resource_with_shapefile_esri_format_cleaned = pu._prepare_resource_format(
+            resource_with_shapefile_esri_format.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('SHAPEFILE', resource_with_shapefile_esri_format_cleaned['format'])
 
         resource_with_text_format = {
@@ -190,8 +209,8 @@ class TestPlugin(unittest.TestCase):
             'media_type': None,
             'format': 'text (.txt)'
         }
-        resource_with_text_format_cleaned = ogdch_language_plugin._prepare_resource_format(
-            resource_with_text_format.copy())
+        resource_with_text_format_cleaned = pu._prepare_resource_format(
+            resource_with_text_format.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('TXT', resource_with_text_format_cleaned['format'])
 
         resource_with_comma_format = {
@@ -199,6 +218,6 @@ class TestPlugin(unittest.TestCase):
             'media_type': None,
             'format': 'comma ...'
         }
-        resource_with_comma_format_cleaned = ogdch_language_plugin._prepare_resource_format(
-            resource_with_comma_format.copy())
+        resource_with_comma_format_cleaned = pu._prepare_resource_format(
+            resource_with_comma_format.copy(), format_mapping=ogdch_language_plugin.format_mapping)
         self.assertEquals('CSV', resource_with_comma_format_cleaned['format'])
